add ehador to layout, create home and ingredients page
