+++ undercooked/src/Pages/Layout/Layout.css
@@ -1,0 +1,32 @@
+nav {
+  margin: 0;
+  padding: 0;
+  list-style-type: none;
+  display: flex;
+  width: 100%;
+  justify-content: flex-start;
+  align-items: center;
+  padding: 0.5rem;
+  color: white;
+  font-size: 1.5rem;
+  font-weight: bold;
+  text-decoration: none;
+  background-color: green;
+  box-shadow: 5px 5px 5px #888888;
+}
+
+.Layout{
+  width: 100%
+}
+
+img{
+  height: 70px;
+  margin-right: 10px;
+}
+
+a{
+  color: #8BC53F;
+  font-size: 70px;
+  text-decoration: none; 
+  font-family:Arial, Helvetica, sans-serif
+}
