+++ undercooked/src/index.css
@@ -1,4 +1,4 @@
-:root {
+/*:root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
@@ -65,4 +65,14 @@ button:focus-visible {
   button {
     background-color: #f9f9f9;
   }
+}*/
+
+body {
+  margin: 0;
+  padding: 0;
+}
+
+.root {
+  margin: 0;
+  padding: 0;
 }
