+++ undercooked/src/main.jsx
@@ -3,12 +3,24 @@ import ReactDOM from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import './index.css'
 
-import Layout from "./Pages/Layout";
+import Layout from "./Pages/Layout/Layout";
+import Home from './Pages/Home/Home';
+import Ingredients from './Pages/Ingredients/Ingredients';
 
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <Layout />
+    element: <Layout />,
+    children: [
+      {
+        path: '/',
+        element: <Home />
+      },
+      {
+        path: '/ingredients',
+        element: <Ingredients />
+      },
+    ]
   }
 ])
